+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/PermissionFilter.java
@@ -1,5 +1,7 @@
 package com.dwarfeng.rbacds.impl.handler;
 
+import com.dwarfeng.rbacds.stack.bean.entity.Permission;
+
 /**
  * 权限过滤器。
  *
@@ -18,9 +20,10 @@ public interface PermissionFilter {
     /**
      * 判断是否接受指定的权限。
      *
-     * @param pattern           权限表达式的匹配模式。
-     * @param permissionContent 权限的内容。
+     * @param pattern    权限表达式的匹配模式。
+     * @param permission 权限实体。
      * @return 是否接受指定的权限。
+     * @throws Exception 测试过程中发生的任何异常。
      */
-    boolean accept(String pattern, String permissionContent);
+    boolean accept(String pattern, Permission permission) throws Exception;
 }