+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/preset/NameRegexPermissionFilter.java
@@ -4,24 +4,22 @@ import com.dwarfeng.rbacds.impl.handler.PermissionFilter;
 import com.dwarfeng.rbacds.stack.bean.entity.Permission;
 import org.springframework.stereotype.Component;
 
-import java.util.Objects;
-
 /**
- * 精确匹配的权限过滤器。
+ * NAME 正则匹配的权限过滤器。
  *
  * @author DwArFeng
  * @since 1.0.0
  */
 @Component
-public class ExactPermissionFilter implements PermissionFilter {
+public class NameRegexPermissionFilter implements PermissionFilter {
 
     @Override
     public String getIdentifier() {
-        return "EXACT";
+        return "NAME_REGEX";
     }
 
     @Override
     public boolean accept(String pattern, Permission permission) {
-        return Objects.equals(pattern, permission.getKey().getStringId());
+        return permission.getName().matches(pattern);
     }
 }